Only delete sqlite datafile if we produced it

+++ scripts/aws/client.py
@@ -338,8 +338,9 @@ class solverThread (threading.Thread):
 
         os.unlink(self.get_stdout_fname() + ".gz")
         os.unlink(self.get_stderr_fname() + ".gz")
-        os.unlink(self.get_sqlite_fname() + ".gz")
         # os.unlink(self.get_perf_fname() + ".gz")
+        if "cryptominisat" in self.indata["solver"]:
+            os.unlink(self.get_sqlite_fname() + ".gz")
 
         return toreturn
